backend/requirements.txt
django>=5.0.2
djangorestframework>=3.9.3
celery>=5.3.5
Pillow>=10.1.0
python-decouple>=3.8
django-cors-headers>=3.10.0 
